wip: before rename symbol
