+++ libs/modules/library/src/LibraryOverview.fragments.graphql
@@ -1,0 +1,4 @@
+fragment GetLibraryOverview__Library on library {
+  id
+  name
+}
