--- libs/modules/library/src/getLibraries/GetLibraries.graphql
@@ -1,5 +1,0 @@
-query GetLibraries {
-  library {
-    ...__Library
-  }
-}
