+++ libs/modules/library/src/libraryExplorer/LibraryExplorer.graphql
@@ -1,0 +1,5 @@
+query GetLibrariesOverview {
+  library {
+    ...GetLibraryOverview__Library
+  }
+}
